Adding the events button
